+++ android/CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] - 2021-03-13
+## [1.1.1] - 2021-03-21
+
+### Added
+
+- Add custom offline screen.
+
+### Fixed
+
+- Fix bug of correct route not opening when clicking a push notification on quit state.
+
+## [1.1.0] - 2021-03-20
 
 ### Added